+++ packages/substrate-bindings/src/codecs/metadata/index.ts
@@ -1,3 +1,4 @@
 export { metadata, decAnyMetadata } from "./metadata"
 export * from "./v14"
 export * from "./v15"
+export { lookup as v14Lookup } from "./lookup"